+++ src/activitypub/identifiers.rs
@@ -61,6 +61,10 @@ pub fn local_object_id(instance_url: &str, internal_object_id: &Uuid) -> String 
     format!("{}/objects/{}", instance_url, internal_object_id)
 }
 
+pub fn local_emoji_id(instance_url: &str, emoji_name: &str) -> String {
+    format!("{}/objects/emojis/{}", instance_url, emoji_name)
+}
+
 pub fn parse_local_actor_id(
     instance_url: &str,
     actor_id: &str,